feat: move metadata to a separate file for better organization
