+++ app/layout.tsx
@@ -1,7 +1,7 @@
 import "./globals.css";
-import type { Metadata } from "next";
 import { Poppins } from "next/font/google";
 import { Provider } from "./providers";
+import { metadata } from "./metadata";
 
 const poppins = Poppins({
   variable: "--font-poppins",
@@ -9,211 +9,7 @@ const poppins = Poppins({
   weight: ["400", "500", "600", "700", "800", "900"],
 });
 
-export const metadata: Metadata = {
-  title: "ChipIn | Simplify Shared Expenses",
-  description:
-    "Effortlessly split costs for trips, events, and daily expenses with our easy-to-use web-based tool.",
-  appleWebApp: {
-    capable: true,
-    title: "ChipIn | Simplify Shared Expenses",
-    statusBarStyle: "default",
-    startupImage: [
-      {
-        media:
-          "(device-width: 1024px) and (device-height: 1366px) and (-webkit-device-pixel-ratio: 2) and (orientation: portrait)",
-        url: "/apple-splash-2048-2732.jpg",
-      },
-      {
-        media:
-          "(device-width: 1024px) and (device-height: 1366px) and (-webkit-device-pixel-ratio: 2) and (orientation: landscape)",
-        url: "/apple-splash-2732-2048.jpg",
-      },
-      {
-        media:
-          "(device-width: 834px) and (device-height: 1194px) and (-webkit-device-pixel-ratio: 2) and (orientation: portrait)",
-        url: "/apple-splash-1668-2388.jpg",
-      },
-      {
-        media:
-          "(device-width: 834px) and (device-height: 1194px) and (-webkit-device-pixel-ratio: 2) and (orientation: landscape)",
-        url: "/apple-splash-2388-1668.jpg",
-      },
-      {
-        media:
-          "(device-width: 768px) and (device-height: 1024px) and (-webkit-device-pixel-ratio: 2) and (orientation: portrait)",
-        url: "/apple-splash-1536-2048.jpg",
-      },
-      {
-        media:
-          "(device-width: 768px) and (device-height: 1024px) and (-webkit-device-pixel-ratio: 2) and (orientation: landscape)",
-        url: "/apple-splash-2048-1536.jpg",
-      },
-      {
-        media:
-          "(device-width: 820px) and (device-height: 1180px) and (-webkit-device-pixel-ratio: 2) and (orientation: portrait)",
-        url: "/apple-splash-1640-2360.jpg",
-      },
-      {
-        media:
-          "(device-width: 820px) and (device-height: 1180px) and (-webkit-device-pixel-ratio: 2) and (orientation: landscape)",
-        url: "/apple-splash-2360-1640.jpg",
-      },
-      {
-        media:
-          "(device-width: 834px) and (device-height: 1112px) and (-webkit-device-pixel-ratio: 2) and (orientation: portrait)",
-        url: "/apple-splash-1668-2224.jpg",
-      },
-      {
-        media:
-          "(device-width: 834px) and (device-height: 1112px) and (-webkit-device-pixel-ratio: 2) and (orientation: landscape)",
-        url: "/apple-splash-2224-1668.jpg",
-      },
-      {
-        media:
-          "(device-width: 810px) and (device-height: 1080px) and (-webkit-device-pixel-ratio: 2) and (orientation: portrait)",
-        url: "/apple-splash-1620-2160.jpg",
-      },
-      {
-        media:
-          "(device-width: 810px) and (device-height: 1080px) and (-webkit-device-pixel-ratio: 2) and (orientation: landscape)",
-        url: "/apple-splash-2160-1620.jpg",
-      },
-      {
-        media:
-          "(device-width: 744px) and (device-height: 1133px) and (-webkit-device-pixel-ratio: 2) and (orientation: portrait)",
-        url: "/apple-splash-1488-2266.jpg",
-      },
-      {
-        media:
-          "(device-width: 744px) and (device-height: 1133px) and (-webkit-device-pixel-ratio: 2) and (orientation: landscape)",
-        url: "/apple-splash-2266-1488.jpg",
-      },
-      {
-        media:
-          "(device-width: 440px) and (device-height: 956px) and (-webkit-device-pixel-ratio: 3) and (orientation: portrait)",
-        url: "/apple-splash-1320-2868.jpg",
-      },
-      {
-        media:
-          "(device-width: 440px) and (device-height: 956px) and (-webkit-device-pixel-ratio: 3) and (orientation: landscape)",
-        url: "/apple-splash-2868-1320.jpg",
-      },
-      {
-        media:
-          "(device-width: 402px) and (device-height: 874px) and (-webkit-device-pixel-ratio: 3) and (orientation: portrait)",
-        url: "/apple-splash-1206-2622.jpg",
-      },
-      {
-        media:
-          "(device-width: 402px) and (device-height: 874px) and (-webkit-device-pixel-ratio: 3) and (orientation: landscape)",
-        url: "/apple-splash-2622-1206.jpg",
-      },
-      {
-        media:
-          "(device-width: 430px) and (device-height: 932px) and (-webkit-device-pixel-ratio: 3) and (orientation: portrait)",
-        url: "/apple-splash-1290-2796.jpg",
-      },
-      {
-        media:
-          "(device-width: 430px) and (device-height: 932px) and (-webkit-device-pixel-ratio: 3) and (orientation: landscape)",
-        url: "/apple-splash-2796-1290.jpg",
-      },
-      {
-        media:
-          "(device-width: 393px) and (device-height: 852px) and (-webkit-device-pixel-ratio: 3) and (orientation: portrait)",
-        url: "/apple-splash-1179-2556.jpg",
-      },
-      {
-        media:
-          "(device-width: 393px) and (device-height: 852px) and (-webkit-device-pixel-ratio: 3) and (orientation: landscape)",
-        url: "/apple-splash-2556-1179.jpg",
-      },
-      {
-        media:
-          "(device-width: 390px) and (device-height: 844px) and (-webkit-device-pixel-ratio: 3) and (orientation: portrait)",
-        url: "/apple-splash-1170-2532.jpg",
-      },
-      {
-        media:
-          "(device-width: 390px) and (device-height: 844px) and (-webkit-device-pixel-ratio: 3) and (orientation: landscape)",
-        url: "/apple-splash-2532-1170.jpg",
-      },
-      {
-        media:
-          "(device-width: 428px) and (device-height: 926px) and (-webkit-device-pixel-ratio: 3) and (orientation: portrait)",
-        url: "/apple-splash-1284-2778.jpg",
-      },
-      {
-        media:
-          "(device-width: 428px) and (device-height: 926px) and (-webkit-device-pixel-ratio: 3) and (orientation: landscape)",
-        url: "/apple-splash-2778-1284.jpg",
-      },
-      {
-        media:
-          "(device-width: 375px) and (device-height: 812px) and (-webkit-device-pixel-ratio: 3) and (orientation: portrait)",
-        url: "/apple-splash-1125-2436.jpg",
-      },
-      {
-        media:
-          "(device-width: 375px) and (device-height: 812px) and (-webkit-device-pixel-ratio: 3) and (orientation: landscape)",
-        url: "/apple-splash-2436-1125.jpg",
-      },
-      {
-        media:
-          "(device-width: 414px) and (device-height: 896px) and (-webkit-device-pixel-ratio: 3) and (orientation: portrait)",
-        url: "/apple-splash-1242-2688.jpg",
-      },
-      {
-        media:
-          "(device-width: 414px) and (device-height: 896px) and (-webkit-device-pixel-ratio: 3) and (orientation: landscape)",
-        url: "/apple-splash-2688-1242.jpg",
-      },
-      {
-        media:
-          "(device-width: 414px) and (device-height: 896px) and (-webkit-device-pixel-ratio: 2) and (orientation: portrait)",
-        url: "/apple-splash-828-1792.jpg",
-      },
-      {
-        media:
-          "(device-width: 414px) and (device-height: 896px) and (-webkit-device-pixel-ratio: 2) and (orientation: landscape)",
-        url: "/apple-splash-1792-828.jpg",
-      },
-      {
-        media:
-          "(device-width: 414px) and (device-height: 736px) and (-webkit-device-pixel-ratio: 3) and (orientation: portrait)",
-        url: "/apple-splash-1242-2208.jpg",
-      },
-      {
-        media:
-          "(device-width: 414px) and (device-height: 736px) and (-webkit-device-pixel-ratio: 3) and (orientation: landscape)",
-        url: "/apple-splash-2208-1242.jpg",
-      },
-      {
-        media:
-          "(device-width: 375px) and (device-height: 667px) and (-webkit-device-pixel-ratio: 2) and (orientation: portrait)",
-        url: "/apple-splash-750-1334.jpg",
-      },
-      {
-        media:
-          "(device-width: 375px) and (device-height: 667px) and (-webkit-device-pixel-ratio: 2) and (orientation: landscape)",
-        url: "/apple-splash-1334-750.jpg",
-      },
-      {
-        media:
-          "(device-width: 320px) and (device-height: 568px) and (-webkit-device-pixel-ratio: 2) and (orientation: portrait)",
-        url: "/apple-splash-640-1136.jpg",
-      },
-      {
-        media:
-          "(device-width: 320px) and (device-height: 568px) and (-webkit-device-pixel-ratio: 2) and (orientation: landscape)",
-        url: "/apple-splash-1136-640.jpg",
-      },
-    ],
-  },
-  other: {
-    "apple-mobile-web-app-capable": "yes",
-  },
-};
+export { metadata };
 
 export default function RootLayout({
   children,
